change: sketch-core module multiplatform developing

+++ sketch-core/src/commonMain/kotlin/com/github/panpf/sketch/cache/internal/KeyMapperCache.kt
@@ -17,7 +17,7 @@ package com.github.panpf.sketch.cache.internal
 
 import com.github.panpf.sketch.util.LruCache
 
-class KeyMapperCache(val maxSize: Int = 100, val mapper: (key: String) -> String) {
+class KeyMapperCache(val maxSize: Long = 100L, val mapper: (key: String) -> String) {
 
     private val cache = LruCache<String, String>(maxSize)
